requirements.txt
boto3==1.26.138
botocore==1.29.138
jmespath==1.0.1
lz4==4.3.2
mcap==1.0.2
numpy==1.23.1
pandas==2.0.1
psycopg2-binary==2.9.6
python-dateutil==2.8.2
pytz==2023.3
rosbags==0.9.15
ruamel.yaml==0.17.26
ruamel.yaml.clib==0.2.7
s3transfer==0.6.1
six==1.16.0
tzdata==2023.3
urllib3==1.26.15
zstandard==0.21.0
